moving JDMapper.kt (<- ToLocalDate.kt) and JDRepository.kt files from the `data` module to the `features:jumps:data` module and refactoring the JDRepository
